requirements.txt
flask==2.0.1
requests==2.28.1
werkzeug==2.0.3
python-dotenv==1.0.0 